Initial commit of Sociomantic's version
This is a one-off commit of changes that happen during more than 5 years
in a fork that we never thought it will see the light. Unfortunately
even at which exact version this was forked is uncertain, and specially
because of the moving of code from headers to C files, it's particularly
hard to track what really changed.

Because of this, this version is just being committed as a complete
replacement of the current repository contents, although we still want
to make it clear that this is a derived work from this library.

The license change (to MIT/X11) was authorized by Willy Tarreau and is
the same license a new experimental branch of the library will use.